requirements.txt
aioredis==1.3.1
amqp==2.5.2
asgiref==3.2.7
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.3.1
Automat==20.2.0
billiard==3.6.3.0
boto3==1.12.28
botocore==1.15.28
celery==4.4.2
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
Django==3.0.4
django-cors-headers==3.2.1
django-hosts==4.0
django-lifecycle==0.7.1
django-rest-framework==0.1.0
django-storages==1.9.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.5.0
incremental==17.5.0
isort==4.3.21
jmespath==0.9.5
kombu==4.6.8
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==0.6.2
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycrypto==2.6.1
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.4.4
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2019.3
redis==3.4.1
s3transfer==0.3.3
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.3.1
typed-ast==1.4.1
urllib3==1.25.8
vine==1.3.0
wrapt==1.11.2
zipp==3.1.0
zope.interface==5.0.1
